+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Made ability for AC, Touch AC and CMD configurable (thanks mkahvi!)
 - Implemented a rolldata variable for spellbooks' ability modifiers, using `@spells.primary.abilityMod` for example (thanks mkahvi!)
 - Added Attack of Opportunity and Nonaction to the list of available actions for items (thanks mkahvi!)
+- Slightly changed style of attack and effect note input
 
 ### Compendiums